% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kage.R
\name{kage}
\alias{kage}
\title{Get Korean Age or Global Age}
\usage{
kage(x, y, z)
}
\arguments{
\item{x}{A birthday, can be any date in a form of year-month-day.}

\item{y}{Any date in a form of year-month-day.}

\item{z}{K for Korean age. G for global age.}
}
\value{
Korean or global age for given two dates: \code{x} and \code{y}.
}
\description{
Get Korean Age or Global Age
}
\examples{
kage("1996-05-28", "2019-10-17", "K")
kage("1992-04-18", "2005-03-04", "G")
}
